Update scripts and xsd2pgschema

+++ lucene_shard.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+sync_update=true
+
 PREFIX=bmr
 ATOM=atom
 
@@ -54,12 +56,12 @@ fi
 
 if [ $ATOM = "atom" ] ; then
 
- XML_RAW_DIR=$PREFIX"_xml_raw"
+ XML_DOC_DIR=$PREFIX"_xml_doc"
  FILE_EXT_DIGEST=.
 
 else
 
- XML_RAW_DIR=$PREFIX"_xml_noatom_raw"
+ XML_DOC_DIR=$PREFIX"_xml_noatom_doc"
  FILE_EXT_DIGEST=-noatom
 
 fi
@@ -81,11 +83,11 @@ if [ -d $IDX_DIR ] ; then
    exit 1;;
  esac
 
- rm -rf $IDX_DIR
+ if [ $sync_update != "true" ] ; then
+  rm -rf $IDX_DIR
+ fi
 
 fi
-
-./$PREFIX"unzip_xml.sh" -a $ATOM
 
 WORK_DIR=lucene_work
 ERR_DIR=$WORK_DIR/err
@@ -93,12 +95,23 @@ ERR_DIR=$WORK_DIR/err
 rm -rf $WORK_DIR
 
 mkdir -p $WORK_DIR
-mkdir -p $IDX_DIR
 mkdir -p $ERR_DIR
+
+if [ $sync_update = "true" ] ; then
+ MD5_DIR=chk_sum_lucene
+fi
 
 err_file=$ERR_DIR/all_err
 
-java -cp extlibs/xsd2pgschema.jar xml2luceneidx --xsd $XSD_SCHEMA --xml $XML_RAW_DIR --idx-dir $IDX_DIR --attr-all --no-rel --no-valid --xml-file-ext-digest $FILE_EXT_DIGEST --shard-size $SHARDS 2> $err_file
+if [ $sync_update != "true" ] ; then
+
+ java -cp extlibs/xsd2pgschema.jar xml2luceneidx --xsd $XSD_SCHEMA --xml $XML_DOC_DIR --idx-dir $IDX_DIR --attr-all --no-rel --no-valid --xml-file-ext gz --xml-file-ext-digest $FILE_EXT_DIGEST --shard-size $SHARDS 2> $err_file
+
+else
+
+ java -cp extlibs/xsd2pgschema.jar xml2luceneidx --xsd $XSD_SCHEMA --xml $XML_DOC_DIR --idx-dir $IDX_DIR --attr-all --no-rel --no-valid --xml-file-ext gz --xml-file-ext-digest $FILE_EXT_DIGEST --shard-size $SHARDS --sync $MD5_DIR 2> $err_file
+
+fi
 
 if [ $? = 0 ] && [ ! -s $err_file ] ; then
  rm -f $err_file
@@ -122,7 +135,6 @@ if [ $errs = 0 ] ; then
    echo "Lucene index (prefix:"$PREFIX"-"$ATOM") is update."
   fi
 
-#  rm -rf $XML_RAW_DIR
   rm -rf $WORK_DIR
 
  fi
